Fixed zipfile typo and dhus max records

+++ iq_download.py
@@ -331,14 +331,15 @@ if options.hub is None:
 elif options.hub == 'dhus':
     huburl = 'https://scihub.copernicus.eu/dhus/'
     # The data hub had a limit of 10 records.
-    options.MaxRecords = '10'
+    if int(options.MaxRecords) > 10:
+        options.MaxRecords = '10'
 # # Untested
 # elif options.hub == 'zamg':
 #     huburl = 'https://data.sentinel.zamg.ac.at/api/'
 
 # Make sure write_dir is formatted properly.
 options.write_dir = (options.write_dir).replace('/', '\\')
-print options.write_dir
+print '\nDirectory: {}\n'.format(options.write_dir)
 
 # Build in checks for valid commands related to the spatial aspect.
 if options.tile is None or options.tile == '?':
@@ -589,7 +590,7 @@ total_size = '{0:.2f} GB'.format(total_size)
 if options.tile is None or options.tile == '?':
     question_tile = 'Do you want to download all results?'
 elif options.tile != None:
-    question_tile = ('Do you want to download only {} tiles selected'
+    question_tile = ('Do you want to download only {} tiles selected '
         'from the results?').format(options.tile)
 
 # Create question to continue based on the number of scenes found.
@@ -641,7 +642,7 @@ if messagebox and (options.tile is None or options.tile == '?'):
             print 'Unzipped Scene # {}'.format(str(entry + 1))
         # If the unzipped and zipped version exist, delete the zipped version.
         if (os.path.exists(os.path.join(options.write_dir, filename))
-                and os.path.exists(os.path.join(options.write_dir, zfile)):
+                and os.path.exists(os.path.join(options.write_dir, zfile))):
             os.remove(os.path.join(options.write_dir, zfile))
 
     print '\n------------------------------------------------------------------'
@@ -750,7 +751,7 @@ elif messagebox and options.tile != None and options.tile != '?':
 
             # Delete the zipped version.
             if (os.path.exists(os.path.join(options.write_dir, filename))
-                    and os.path.exists(os.path.join(options.write_dir, zfile)):
+                    and os.path.exists(os.path.join(options.write_dir, zfile))):
                 os.remove(os.path.join(options.write_dir, zfile))
 
         else:
